Move SQL functions to the sql.go file.
