+++ wrapper.go
@@ -22,17 +22,13 @@
 package db
 
 import (
-	"database/sql"
 	"fmt"
 	"sync"
 )
 
 var (
-	sqlAdapters map[string]*SQLAdapterFuncMap
-	adapters    map[string]*AdapterFuncMap
-
-	sqlAdaptersMu sync.RWMutex
-	adaptersMu    sync.RWMutex
+	adapters   map[string]*AdapterFuncMap
+	adaptersMu sync.RWMutex
 )
 
 func init() {
@@ -44,12 +40,8 @@ type AdapterFuncMap struct {
 	Open func(settings ConnectionURL) (Database, error)
 }
 
-type SQLAdapterFuncMap struct {
-	New   func(sqlDB *sql.DB) (SQLDatabase, error)
-	NewTx func(sqlTx *sql.Tx) (SQLTx, error)
-	Open  func(settings ConnectionURL) (SQLDatabase, error)
-}
-
+// RegisterAdapter registers a generic Database adapter. This function must be
+// called from adapter packages upon initialization.
 func RegisterAdapter(name string, adapter *AdapterFuncMap) {
 	adaptersMu.Lock()
 	defer adaptersMu.Unlock()
@@ -63,25 +55,7 @@ func RegisterAdapter(name string, adapter *AdapterFuncMap) {
 	adapters[name] = adapter
 }
 
-func RegisterSQLAdapter(name string, adapter *SQLAdapterFuncMap) {
-	sqlAdaptersMu.Lock()
-	defer sqlAdaptersMu.Unlock()
-
-	if name == "" {
-		panic(`Missing adapter name`)
-	}
-	if _, ok := sqlAdapters[name]; ok {
-		panic(`db.RegisterSQLAdapter() called twice for adapter: ` + name)
-	}
-	sqlAdapters[name] = adapter
-
-	RegisterAdapter(name, &AdapterFuncMap{
-		Open: func(settings ConnectionURL) (Database, error) {
-			return adapter.Open(settings)
-		},
-	})
-}
-
+// Adapter returns a method map from the given adapter.
 func Adapter(name string) AdapterFuncMap {
 	adaptersMu.RLock()
 	defer adaptersMu.RUnlock()
@@ -92,35 +66,10 @@ func Adapter(name string) AdapterFuncMap {
 	return missingAdapter(name)
 }
 
-func SQLAdapter(name string) SQLAdapterFuncMap {
-	sqlAdaptersMu.RLock()
-	defer sqlAdaptersMu.RUnlock()
-
-	if fn, ok := sqlAdapters[name]; ok {
-		return *fn
-	}
-	return missingSQLAdapter(name)
-}
-
 func missingAdapter(name string) AdapterFuncMap {
 	err := fmt.Errorf("upper: Missing adapter %q, forgot to import?", name)
 	return AdapterFuncMap{
 		Open: func(ConnectionURL) (Database, error) {
-			return nil, err
-		},
-	}
-}
-
-func missingSQLAdapter(name string) SQLAdapterFuncMap {
-	err := fmt.Errorf("upper: Missing SQL adapter %q, forgot to import?", name)
-	return SQLAdapterFuncMap{
-		New: func(*sql.DB) (SQLDatabase, error) {
-			return nil, err
-		},
-		NewTx: func(*sql.Tx) (SQLTx, error) {
-			return nil, err
-		},
-		Open: func(ConnectionURL) (SQLDatabase, error) {
 			return nil, err
 		},
 	}
